pyproject.toml
[project]
name = "new-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Olga",email = "tasenko_03@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "torchaudio (>=2.7.0,<3.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "webrtcvad (>=2.0.10,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "pesq (>=0.0.4,<0.0.5)",
    "librosa (>=0.11.0,<0.12.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "jupyterlab (>=4.4.2,<5.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "speechbrain (>=1.0.3,<2.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "openai-whisper (>=20240930,<20240931)",
    "praat-parselmouth (>=0.4.5,<0.5.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "natasha (>=1.6.0,<2.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "textstat (>=0.7.7,<0.8.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "razdel (>=0.5.0,<0.6.0)",
    "pymorphy2 (>=0.9.1,<0.10.0)",
    "pymorphy2-dicts-ru (>=2.4.417127.4579844,<3.0.0.0)",
    "stanza (>=1.10.1,<2.0.0)",
    "peft (>=0.15.2,<0.16.0)",
    "requests (>=2.32.3,<3.0.0)",
    "lexical-diversity (>=0.1.1,<0.2.0)",
    "spicy (>=0.16.0,<0.17.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "pymorphy3 (>=2.0.3,<3.0.0)",
    "prettytable (>=3.16.0,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pydrive (>=1.3.1,<2.0.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "xgboost (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
